+++ packages/workspace-store/src/schemas/reference-config/settings.ts
@@ -1,5 +1,5 @@
 import { Type } from '@scalar/typebox'
-import type { RequiredDeep } from 'type-fest/source/required-deep'
+import type { RequiredDeep } from 'type-fest'
 
 import { ServerObjectSchema } from '@/schemas/v3.1/strict/openapi-document'
 import type { ServerObject } from '@/schemas/v3.1/strict/server'