+++ integrations/nestjs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @scalar/nestjs-api-reference
 
+## 1.0.5
+
+### Patch Changes
+
+- @scalar/core@0.3.20
+
 ## 1.0.4
 
 ### Patch Changes