+++ packages/oas-utils/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @scalar/oas-utils
 
+## 0.5.2
+
+### Patch Changes
+
+- f8efecd: feat: use new schema example generation in the response
+- Updated dependencies [3f6d0b9]
+- Updated dependencies [2089748]
+- Updated dependencies [f3e17d8]
+- Updated dependencies [8a7fb2a]
+- Updated dependencies [1335923]
+- Updated dependencies [07397c8]
+- Updated dependencies [f69e7cc]
+  - @scalar/helpers@0.0.12
+  - @scalar/json-magic@0.6.1
+  - @scalar/types@0.3.2
+  - @scalar/workspace-store@0.17.1
+  - @scalar/openapi-types@0.5.0
+  - @scalar/object-utils@1.2.8
+  - @scalar/themes@0.13.22
+
 ## 0.5.1
 
 ### Patch Changes